@ Upgrade to Spring Boot 3.4.5 + Java 17, add Spring AI agent, RAG pipeline, Elasticsearch search, and rate limiting
- Upgrade Spring Boot 2.3 → 3.4.5, Java 8 → 17, MySQL 5.7 → 8.0, Redis → Redis Stack 7.2
- Add Spring AI-powered recommendation agent with 5 tools, chat memory, and preference extraction
- Add configurable RAG pipeline: MQE + HyDE → multi-query retrieval → LLM rerank + sentiment tagging
- Add Elasticsearch full-text shop search with geo-distance sorting and result highlighting
- Add declarative @RateLimit annotation with Redisson RRateLimiter
- Add SpringDoc OpenAPI (Swagger UI)
- Add GitHub Actions CI workflow
- Update Docker Compose with Elasticsearch and Redis Stack
- Update README (EN/CN) with complete feature documentation
@

+++ src/main/java/com/citylife/service/impl/FollowServiceImpl.java
@@ -13,7 +13,7 @@ import com.citylife.utils.UserHolder;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
@@ -56,7 +56,7 @@ public class FollowServiceImpl extends ServiceImpl<FollowMapper, Follow> impleme
     @Override
     public Result<Boolean> isFollow(Long followUserId) {
         Long userId = UserHolder.getUser().getId();
-        Integer count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
+        Long count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
         return Result.ok(count > 0);
     }
